src/contextual-signal-encoder/requirements.txt
fastapi>=0.115,<1
uvicorn[standard]>=0.30,<1
pydantic>=2.0,<3
pydantic-settings>=2.0,<3
numpy>=1.26,<3
httpx>=0.27,<1
sentence-transformers>=3.0,<4
beautifulsoup4>=4.12,<5
pyyaml>=6.0,<7
pytest>=8.0,<9
pytest-asyncio>=0.23,<1
